+++ singer_sdk/connectors/__init__.py
@@ -1,0 +1,5 @@
+"""Module for SQL-related operations."""
+
+from .sql import SQLConnector
+
+__all__ = ["SQLConnector"]