+++ singer_sdk/sinks/sql.py
@@ -1,5 +1,8 @@
 """Sink classes load data to SQL targets."""
 
+import re
+from collections import defaultdict
+from copy import copy
 from textwrap import dedent
 from typing import Any, Dict, Iterable, List, Optional, Type, Union
 
@@ -8,9 +11,11 @@ from pendulum import now
 from sqlalchemy.sql import Executable
 from sqlalchemy.sql.expression import bindparam
 
+from singer_sdk.connectors import SQLConnector
+from singer_sdk.exceptions import ConformedNameClashException
+from singer_sdk.helpers._conformers import replace_leading_digit, snakecase
 from singer_sdk.plugin_base import PluginBase
 from singer_sdk.sinks.batch import BatchSink
-from singer_sdk.streams.sql import SQLConnector
 
 
 class SQLSink(BatchSink):
@@ -38,11 +43,7 @@ class SQLSink(BatchSink):
             connector: Optional connector to reuse.
         """
         self._connector: SQLConnector
-        if connector:
-            self._connector = connector
-        else:
-            self._connector = self.connector_class(dict(target.config))
-
+        self._connector = connector or self.connector_class(dict(target.config))
         super().__init__(target, stream_name, schema, key_properties)
 
     @property
@@ -65,35 +66,167 @@ class SQLSink(BatchSink):
 
     @property
     def table_name(self) -> str:
-        """Returns the table name, with no schema or database part.
+        """Return the table name, with no schema or database part.
 
         Returns:
             The target table name.
         """
         parts = self.stream_name.split("-")
-
-        if len(parts) == 1:
-            return self.stream_name
-        else:
-            return parts[-1]
+        table = self.stream_name if len(parts) == 1 else parts[-1]
+        return self.conform_name(table, "table")
 
     @property
     def schema_name(self) -> Optional[str]:
-        """Returns the schema name or `None` if using names with no schema part.
+        """Return the schema name or `None` if using names with no schema part.
 
         Returns:
             The target schema name.
         """
-        return None  # Assumes single-schema target context.
+        parts = self.stream_name.split("-")
+        if len(parts) in {2, 3}:
+            # Stream name is a two-part or three-part identifier.
+            # Use the second-to-last part as the schema name.
+            return self.conform_name(parts[-2], "schema")
+
+        # Schema name not detected.
+        return None
 
     @property
     def database_name(self) -> Optional[str]:
-        """Returns the DB name or `None` if using names with no database part.
+        """Return the DB name or `None` if using names with no database part.
 
         Returns:
             The target database name.
         """
         return None  # Assumes single-DB target context.
+
+    @property
+    def full_table_name(self) -> str:
+        """Return the fully qualified table name.
+
+        Returns:
+            The fully qualified table name.
+        """
+        return self.connector.get_fully_qualified_name(
+            table_name=self.table_name,
+            schema_name=self.schema_name,
+            db_name=self.database_name,
+        )
+
+    @property
+    def full_schema_name(self) -> str:
+        """Return the fully qualified schema name.
+
+        Returns:
+            The fully qualified schema name.
+        """
+        return self.connector.get_fully_qualified_name(
+            schema_name=self.schema_name, db_name=self.database_name
+        )
+
+    def conform_name(self, name: str, object_type: Optional[str] = None) -> str:
+        """Conform a stream property name to one suitable for the target system.
+
+        Transforms names to snake case by default, applicable to most common DBMSs'.
+        Developers may override this method to apply custom transformations
+        to database/schema/table/column names.
+
+        Args:
+            name: Property name.
+            object_type: One of ``database``, ``schema``, ``table`` or ``column``.
+
+
+        Returns:
+            The name transformed to snake case.
+        """
+        # strip non-alphanumeric characters, keeping - . _ and spaces
+        name = re.sub(r"[^a-zA-Z0-9_\-\.\s]", "", name)
+        # convert to snakecase
+        name = snakecase(name)
+        # replace leading digit
+        return replace_leading_digit(name)
+
+    @staticmethod
+    def _check_conformed_names_not_duplicated(
+        conformed_property_names: Dict[str, str]
+    ) -> None:
+        """Check if conformed names produce duplicate keys.
+
+        Args:
+            conformed_property_names: A name:conformed_name dict map.
+
+        Raises:
+            ConformedNameClashException: if duplicates found.
+        """
+        # group: {'_a': ['1_a'], 'abc': ['aBc', 'abC']}
+        grouped = defaultdict(list)
+        for k, v in conformed_property_names.items():
+            grouped[v].append(k)
+
+        # filter
+        duplicates = list(filter(lambda p: len(p[1]) > 1, grouped.items()))
+        if duplicates:
+            raise ConformedNameClashException(
+                "Duplicate stream properties produced when "
+                + f"conforming property names: {duplicates}"
+            )
+
+    def conform_schema(self, schema: dict) -> dict:
+        """Return schema dictionary with property names conformed.
+
+        Args:
+            schema: JSON schema dictionary.
+
+        Returns:
+            A schema dictionary with the property names conformed.
+        """
+        conformed_schema = copy(schema)
+        conformed_property_names = {
+            key: self.conform_name(key) for key in conformed_schema["properties"].keys()
+        }
+        self._check_conformed_names_not_duplicated(conformed_property_names)
+        conformed_schema["properties"] = {
+            conformed_property_names[key]: value
+            for key, value in conformed_schema["properties"].items()
+        }
+        return conformed_schema
+
+    def conform_record(self, record: dict) -> dict:
+        """Return record dictionary with property names conformed.
+
+        Args:
+            record: Dictionary representing a single record.
+
+        Returns:
+            New record dictionary with conformed column names.
+        """
+        conformed_property_names = {key: self.conform_name(key) for key in record}
+        self._check_conformed_names_not_duplicated(conformed_property_names)
+        return {conformed_property_names[key]: value for key, value in record.items()}
+
+    def setup(self) -> None:
+        """Set up Sink.
+
+        This method is called on Sink creation, and creates the required Schema and
+        Table entities in the target database.
+        """
+        if self.schema_name:
+            self.connector.prepare_schema(self.schema_name)
+        self.connector.prepare_table(
+            full_table_name=self.full_table_name,
+            schema=self.conform_schema(self.schema),
+            primary_keys=self.key_properties,
+            as_temp_table=False,
+        )
+
+    @property
+    def key_properties(self) -> List[str]:
+        """Return key properties, conformed to target system naming requirements.
+
+        Returns:
+            A list of key properties, conformed with `self.conform_name()`
+        """
+        return [self.conform_name(key, "column") for key in super().key_properties]
 
     def process_batch(self, context: dict) -> None:
         """Process a batch with the given batch context.
@@ -106,62 +239,10 @@ class SQLSink(BatchSink):
         """
         # If duplicates are merged, these can be tracked via
         # :meth:`~singer_sdk.Sink.tally_duplicate_merged()`.
-        self.connector.prepare_table(
-            full_table_name=self.full_table_name,
-            schema=self.schema,
-            primary_keys=self.key_properties,
-            as_temp_table=False,
-        )
         self.bulk_insert_records(
             full_table_name=self.full_table_name,
             schema=self.schema,
             records=context["records"],
-        )
-
-    @property
-    def full_table_name(self) -> str:
-        """Gives the fully qualified table name.
-
-        Returns:
-            The fully qualified table name.
-        """
-        return self.connector.get_fully_qualified_name(
-            self.table_name,
-            self.schema_name,
-            self.database_name,
-        )
-
-    def create_table_with_records(
-        self,
-        full_table_name: Optional[str],
-        schema: dict,
-        records: Iterable[Dict[str, Any]],
-        primary_keys: Optional[List[str]] = None,
-        partition_keys: Optional[List[str]] = None,
-        as_temp_table: bool = False,
-    ) -> None:
-        """Create an empty table.
-
-        Args:
-            full_table_name: the target table name.
-            schema: the JSON schema for the new table.
-            records: records to load.
-            primary_keys: list of key properties.
-            partition_keys: list of partition keys.
-            as_temp_table: True to create a temp table.
-        """
-        full_table_name = full_table_name or self.full_table_name
-        if primary_keys is None:
-            primary_keys = self.key_properties
-        partition_keys = partition_keys or None
-        self.connector.prepare_table(
-            full_table_name=full_table_name,
-            primary_keys=primary_keys,
-            schema=schema,
-            as_temp_table=as_temp_table,
-        )
-        self.bulk_insert_records(
-            full_table_name=full_table_name, schema=schema, records=records
         )
 
     def generate_insert_statement(
@@ -178,7 +259,7 @@ class SQLSink(BatchSink):
         Returns:
             An insert statement.
         """
-        property_names = list(schema["properties"].keys())
+        property_names = list(self.conform_schema(schema)["properties"].keys())
         statement = dedent(
             f"""\
             INSERT INTO {full_table_name}
@@ -186,7 +267,6 @@ class SQLSink(BatchSink):
             VALUES ({", ".join([f":{name}" for name in property_names])})
             """
         )
-
         return statement.rstrip()
 
     def bulk_insert_records(
@@ -217,12 +297,14 @@ class SQLSink(BatchSink):
         if isinstance(insert_sql, str):
             insert_sql = sqlalchemy.text(insert_sql)
 
+        conformed_records = (
+            [self.conform_record(record) for record in records]
+            if isinstance(records, list)
+            else (self.conform_record(record) for record in records)
+        )
         self.logger.info("Inserting with SQL: %s", insert_sql)
-        self.connector.connection.execute(insert_sql, records)
-        if isinstance(records, list):
-            return len(records)  # If list, we can quickly return record count.
-
-        return None  # Unknown record count.
+        self.connector.connection.execute(insert_sql, conformed_records)
+        return len(conformed_records) if isinstance(conformed_records, list) else None
 
     def merge_upsert_from_table(
         self, target_table_name: str, from_table_name: str, join_keys: List[str]