+++ singer_sdk/streams/core.py
@@ -12,13 +12,13 @@ import logging
 from os import PathLike
 from pathlib import Path
 from types import MappingProxyType
-from typing import Any, Callable, Generator, Iterable, Iterator, Mapping, TypeVar, cast
+from typing import Any, Generator, Iterable, Iterator, Mapping, TypeVar, cast
 from uuid import uuid4
 
 import pendulum
-import requests
 
 import singer_sdk._singerlib as singer
+from singer_sdk import metrics
 from singer_sdk.exceptions import InvalidStreamSortException, MaxRecordsLimitException
 from singer_sdk.helpers._batch import (
     BaseBatchFileEncoding,
@@ -51,8 +51,6 @@ REPLICATION_LOG_BASED = "LOG_BASED"
 
 FactoryType = TypeVar("FactoryType", bound="Stream")
 _T = TypeVar("_T")
-
-METRICS_LOG_LEVEL_SETTING = "metrics_log_level"
 
 
 def lazy_chunked_generator(
@@ -116,6 +114,7 @@ class Stream(metaclass=abc.ABCMeta):
             raise ValueError("Missing argument or class variable 'name'.")
 
         self.logger: logging.Logger = tap.logger
+        self.metrics_logger = tap.metrics_logger
         self.tap_name: str = tap.name
         self._config: dict = dict(tap.config)
         self._tap = tap
@@ -696,8 +695,11 @@ class Stream(metaclass=abc.ABCMeta):
     ) -> None:
         """Update state of stream or partition with data from the provided record.
 
-        Raises InvalidStreamSortException is self.is_sorted = True and unsorted data is
-        detected.
+        Raises `InvalidStreamSortException` is `self.is_sorted = True` and unsorted data
+        is detected.
+
+        Note: The default implementation does not advance any bookmarks unless
+        `self.replication_method == 'INCREMENTAL'.
 
         Args:
             latest_record: TODO
@@ -706,28 +708,27 @@ class Stream(metaclass=abc.ABCMeta):
         Raises:
             ValueError: TODO
         """
+        # This also creates a state entry if one does not yet exist:
         state_dict = self.get_context_state(context)
-        if latest_record:
-            if self.replication_method in [
-                REPLICATION_INCREMENTAL,
-                REPLICATION_LOG_BASED,
-            ]:
-                if not self.replication_key:
-                    raise ValueError(
-                        f"Could not detect replication key for '{self.name}' stream"
-                        f"(replication method={self.replication_method})"
-                    )
-                treat_as_sorted = self.is_sorted
-                if not treat_as_sorted and self.state_partitioning_keys is not None:
-                    # Streams with custom state partitioning are not resumable.
-                    treat_as_sorted = False
-                increment_state(
-                    state_dict,
-                    replication_key=self.replication_key,
-                    latest_record=latest_record,
-                    is_sorted=treat_as_sorted,
-                    check_sorted=self.check_sorted,
+
+        # Advance state bookmark values if applicable
+        if latest_record and self.replication_method == REPLICATION_INCREMENTAL:
+            if not self.replication_key:
+                raise ValueError(
+                    f"Could not detect replication key for '{self.name}' stream"
+                    f"(replication method={self.replication_method})"
                 )
+            treat_as_sorted = self.is_sorted
+            if not treat_as_sorted and self.state_partitioning_keys is not None:
+                # Streams with custom state partitioning are not resumable.
+                treat_as_sorted = False
+            increment_state(
+                state_dict,
+                replication_key=self.replication_key,
+                latest_record=latest_record,
+                is_sorted=treat_as_sorted,
+                check_sorted=self.check_sorted,
+            )
 
     # Private message authoring methods:
 
@@ -832,95 +833,13 @@ class Stream(metaclass=abc.ABCMeta):
             )
         )
 
-    @property
-    def _metric_logging_function(self) -> Callable | None:
-        """Return the metrics logging function.
-
-        Returns:
-            The logging function for emitting metrics.
-
-        Raises:
-            ValueError: If logging level setting is an unsupported value.
-        """
-        if METRICS_LOG_LEVEL_SETTING not in self.config:
-            return self.logger.info
-
-        if self.config[METRICS_LOG_LEVEL_SETTING].upper() == "INFO":
-            return self.logger.info
-
-        if self.config[METRICS_LOG_LEVEL_SETTING].upper() == "DEBUG":
-            return self.logger.debug
-
-        if self.config[METRICS_LOG_LEVEL_SETTING].upper() == "NONE":
-            return None
-
-        raise ValueError(
-            "Unexpected logging level for metrics: "
-            + self.config[METRICS_LOG_LEVEL_SETTING]
-        )
-
-    def _write_metric_log(self, metric: dict, extra_tags: dict | None) -> None:
-        """Emit a metric log. Optionally with appended tag info.
+    def _log_metric(self, point: metrics.Point) -> None:
+        """Log a single measurement.
 
         Args:
-            metric: TODO
-            extra_tags: TODO
-
-        Returns:
-            None
+            point: A single measurement value.
         """
-        if not self._metric_logging_function:
-            return None
-
-        if extra_tags:
-            metric["tags"].update(extra_tags)
-        self._metric_logging_function(f"INFO METRIC: {json.dumps(metric)}")
-
-    def _write_record_count_log(self, record_count: int, context: dict | None) -> None:
-        """Emit a metric log. Optionally with appended tag info.
-
-        Args:
-            record_count: TODO
-            context: Stream partition or context dictionary.
-        """
-        extra_tags = {} if not context else {"context": context}
-        counter_metric: dict[str, Any] = {
-            "type": "counter",
-            "metric": "record_count",
-            "value": record_count,
-            "tags": {"stream": self.name},
-        }
-        self._write_metric_log(counter_metric, extra_tags=extra_tags)
-
-    def _write_request_duration_log(
-        self,
-        endpoint: str,
-        response: requests.Response,
-        context: dict | None,
-        extra_tags: dict | None,
-    ) -> None:
-        """TODO.
-
-        Args:
-            endpoint: TODO
-            response: TODO
-            context: Stream partition or context dictionary.
-            extra_tags: TODO
-        """
-        request_duration_metric: dict[str, Any] = {
-            "type": "timer",
-            "metric": "http_request_duration",
-            "value": response.elapsed.total_seconds(),
-            "tags": {
-                "endpoint": endpoint,
-                "http_status_code": response.status_code,
-                "status": "succeeded" if response.status_code < 400 else "failed",
-            },
-        }
-        extra_tags = extra_tags or {}
-        if context:
-            extra_tags["context"] = context
-        self._write_metric_log(metric=request_duration_metric, extra_tags=extra_tags)
+        metrics.log(self.metrics_logger, point=point)
 
     def log_sync_costs(self) -> None:
         """Log a summary of Sync costs.
@@ -1040,70 +959,80 @@ class Stream(metaclass=abc.ABCMeta):
         Yields:
             Each record from the source.
         """
+        # Initialize metrics
+        record_counter = metrics.record_counter(self.name)
+        timer = metrics.sync_timer(self.name)
+
         record_count = 0
         current_context: dict | None
         context_list: list[dict] | None
         context_list = [context] if context is not None else self.partitions
         selected = self.selected
 
-        for current_context in context_list or [{}]:
-            partition_record_count = 0
-            current_context = current_context or None
-            state = self.get_context_state(current_context)
-            state_partition_context = self._get_state_partition_context(current_context)
-            self._write_starting_replication_value(current_context)
-            child_context: dict | None = (
-                None if current_context is None else copy.copy(current_context)
-            )
+        with record_counter, timer:
+            for current_context in context_list or [{}]:
+                record_counter.context = current_context
+                timer.context = current_context
 
-            for record_result in self.get_records(current_context):
-                if isinstance(record_result, tuple):
-                    # Tuple items should be the record and the child context
-                    record, child_context = record_result
-                else:
-                    record = record_result
-                try:
-                    self._process_record(
-                        record,
-                        child_context=child_context,
-                        partition_context=state_partition_context,
-                    )
-                except InvalidStreamSortException as ex:
-                    log_sort_error(
-                        log_fn=self.logger.error,
-                        ex=ex,
-                        record_count=record_count + 1,
-                        partition_record_count=partition_record_count + 1,
-                        current_context=current_context,
-                        state_partition_context=state_partition_context,
-                        stream_name=self.name,
-                    )
-                    raise ex
+                partition_record_count = 0
+                current_context = current_context or None
+                state = self.get_context_state(current_context)
+                state_partition_context = self._get_state_partition_context(
+                    current_context
+                )
+                self._write_starting_replication_value(current_context)
+                child_context: dict | None = (
+                    None if current_context is None else copy.copy(current_context)
+                )
 
-                self._check_max_record_limit(record_count)
+                for record_result in self.get_records(current_context):
+                    if isinstance(record_result, tuple):
+                        # Tuple items should be the record and the child context
+                        record, child_context = record_result
+                    else:
+                        record = record_result
+                    try:
+                        self._process_record(
+                            record,
+                            child_context=child_context,
+                            partition_context=state_partition_context,
+                        )
+                    except InvalidStreamSortException as ex:
+                        log_sort_error(
+                            log_fn=self.logger.error,
+                            ex=ex,
+                            record_count=record_count + 1,
+                            partition_record_count=partition_record_count + 1,
+                            current_context=current_context,
+                            state_partition_context=state_partition_context,
+                            stream_name=self.name,
+                        )
+                        raise ex
 
-                if selected:
-                    if (
-                        record_count - 1
-                    ) % self.STATE_MSG_FREQUENCY == 0 and write_messages:
-                        self._write_state_message()
-                    if write_messages:
-                        self._write_record_message(record)
-                    self._increment_stream_state(record, context=current_context)
+                    self._check_max_record_limit(record_count)
 
-                    yield record
+                    if selected:
+                        if (
+                            record_count - 1
+                        ) % self.STATE_MSG_FREQUENCY == 0 and write_messages:
+                            self._write_state_message()
+                        if write_messages:
+                            self._write_record_message(record)
+                        self._increment_stream_state(record, context=current_context)
 
-                    record_count += 1
-                    partition_record_count += 1
+                        yield record
 
-            if current_context == state_partition_context:
-                # Finalize per-partition state only if 1:1 with context
-                finalize_state_progress_markers(state)
+                        record_counter.increment()
+                        record_count += 1
+                        partition_record_count += 1
+
+                if current_context == state_partition_context:
+                    # Finalize per-partition state only if 1:1 with context
+                    finalize_state_progress_markers(state)
         if not context:
             # Finalize total stream only if we have the full full context.
             # Otherwise will be finalized by tap at end of sync.
             finalize_state_progress_markers(self.stream_state)
-        self._write_record_count_log(record_count=record_count, context=context)
 
         if write_messages:
             # Reset interim bookmarks before emitting final STATE message:
@@ -1120,9 +1049,11 @@ class Stream(metaclass=abc.ABCMeta):
             batch_config: The batch configuration.
             context: Stream partition or context dictionary.
         """
-        for encoding, manifest in self.get_batches(batch_config, context):
-            self._write_batch_message(encoding=encoding, manifest=manifest)
-            self._write_state_message()
+        with metrics.batch_counter(self.name, context=context) as counter:
+            for encoding, manifest in self.get_batches(batch_config, context):
+                counter.increment()
+                self._write_batch_message(encoding=encoding, manifest=manifest)
+                self._write_state_message()
 
     # Public methods ("final", not recommended to be overridden)