+++ tests/core/test_batch.py
@@ -32,9 +32,50 @@ def test_storage_get_url():
         assert url.replace("\\", "/").endswith("root_dir/prefix--file.jsonl.gz")
 
 
-def test_storage_from_url():
-    url = urlparse("s3://bucket/path/to/file?region=us-east-1")
-    target = StorageTarget.from_url(url)
-    assert target.root == "s3://bucket"
-    assert target.prefix is None
-    assert target.params == {"region": ["us-east-1"]}
+@pytest.mark.parametrize(
+    "file_url,root",
+    [
+        pytest.param(
+            "file:///Users/sdk/path/to/file",
+            "file:///Users/sdk/path/to",
+            id="local",
+        ),
+    ],
+)
+def test_storage_from_url(file_url: str, root: str):
+    """Test storage target from URL."""
+    head, _ = StorageTarget.split_url(file_url)
+    target = StorageTarget.from_url(head)
+    assert target.root == root
+
+
+@pytest.mark.parametrize(
+    "file_url,expected",
+    [
+        pytest.param(
+            "file:///Users/sdk/path/to/file",
+            ("file:///Users/sdk/path/to", "file"),
+            id="local",
+        ),
+        pytest.param(
+            "s3://bucket/path/to/file",
+            ("s3://bucket/path/to", "file"),
+            id="s3",
+        ),
+        pytest.param(
+            "file://C:\\Users\\sdk\\path\\to\\file",
+            ("file://C:\\Users\\sdk\\path\\to", "file"),
+            marks=(pytest.mark.windows,),
+            id="windows-local",
+        ),
+        pytest.param(
+            "file://\\\\remotemachine\\C$\\batches\\file",
+            ("file://\\\\remotemachine\\C$\\batches", "file"),
+            marks=(pytest.mark.windows,),
+            id="windows-local",
+        ),
+    ],
+)
+def test_storage_split_url(file_url: str, expected: tuple):
+    """Test storage target split URL."""
+    assert StorageTarget.split_url(file_url) == expected